refactor the emote provider.
