+++ src/main/java/com/karelmikie3/craftcord/resources/CachedEmoteProvider.java
@@ -16,45 +16,28 @@
 
 package com.karelmikie3.craftcord.resources;
 
-import com.google.gson.*;
-import com.karelmikie3.craftcord.CraftCord;
-import com.karelmikie3.craftcord.api.emotes.IEmoteProvider;
-import com.karelmikie3.craftcord.network.RequestEmoteMessageC2S;
+import javafx.util.Pair;
 import net.minecraft.client.Minecraft;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.concurrent.ThreadTaskExecutor;
 import net.minecraftforge.fml.LogicalSide;
 import net.minecraftforge.fml.LogicalSidedProvider;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 
 import java.io.*;
-import java.net.MalformedURLException;
-import java.net.URL;
-import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
-import java.util.*;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
+import java.util.HashSet;
+import java.util.Set;
 
-import static com.karelmikie3.craftcord.resources.MemoryEmoteProvider.*;
-
-public class CachedEmoteProvider implements IEmoteProvider {
+public class CachedEmoteProvider extends AbstractEmoteProvider {
     private static final Minecraft mc = Minecraft.getInstance();
-    private static final Logger LOGGER = LogManager.getLogger();
-    private static final Gson GSON = new GsonBuilder().create();
     private static final Object LOCK = new Object();
-    private static final ExecutorService emoteProcessor = Executors.newFixedThreadPool(2);
 
     private static final File CACHE_LOCATION = new File("assets/craftcordemotes/");
     private static final File EMOTE_LOCATION = new File(CACHE_LOCATION, "emotes");
     private static final File META_LOCATION = new File(CACHE_LOCATION, "metadata");
 
-    //TODO: clear these maps on logout of client.
+    //TODO: clear this set on logout of client.
     private final Set<Long> available = new HashSet<>();
-    private final Map<String, Long> displayNameToID = new HashMap<>();
-    private final Set<String> usable = new HashSet<>();
-    private final Set<Long> requested = new HashSet<>();
 
     private final String URLTemplate;
 
@@ -82,70 +65,28 @@ public class CachedEmoteProvider implements IEmoteProvider {
         }
     }
 
-    @Override
-    public void prepare(long emoteID, String displayName, boolean usable, boolean animated) {
-        emoteProcessor.submit(() -> prepare0(emoteID, displayName, usable, animated));
-    }
-
-    private void prepare0(long emoteID, String displayName, boolean usable, boolean animated) {
-        File emoteFile = new File(EMOTE_LOCATION, Long.toString(emoteID));
-        File metadataFile = new File(META_LOCATION, Long.toString(emoteID));
+    protected void prepareASync(long emoteID, String displayName, boolean usable, boolean animated) {
+        File emoteFile = new File(EMOTE_LOCATION, emoteID + ".png");
+        File metadataFile = new File(META_LOCATION, emoteID + ".mcmeta");
 
         if (!emoteFile.exists() || !metadataFile.exists()) {
             String urlString = String.format(this.URLTemplate, emoteID, animated ? "gif" : "png");
-            URL url;
-            try {
-                url = new URL(urlString);
-            } catch (MalformedURLException e) {
-                LOGGER.fatal("Emote download URL invalid: " + urlString);
-                LOGGER.debug(e);
-                return;
-            }
 
-            byte[] emoteData = download(url);
-            if (emoteData == null)
+            Pair<byte[], byte[]> dataMetadataPair = processEmote(urlString, animated);
+
+            if (dataMetadataPair == null)
                 return;
 
-            JsonObject emoteJson = new JsonObject();
-
-            if (animated) {
-                AnimatedEmoteData data = processAnimatedEmote(emoteData);
-                if (data == null)
-                    return;
-
-                emoteData = data.getEmoteImage();
-                JsonArray delaysJson = new JsonArray();
-                data.getDelays().forEach(delaysJson::add);
-                emoteJson.add("delays", delaysJson);
-                emoteJson.add("height", new JsonPrimitive(data.getDim()));
-                emoteJson.add("frameAmount", new JsonPrimitive(data.getFrameAmount()));
-                emoteJson.add("animated", new JsonPrimitive(true));
-            } else {
-                emoteData = processNormalEmote(emoteData);
-                //TODO: make it so this doesn't have to be.
-                emoteJson.add("delays", new JsonArray());
-                emoteJson.add("height", new JsonPrimitive(-1));
-                emoteJson.add("frameAmount", new JsonPrimitive(-1));
-                emoteJson.add("animated", new JsonPrimitive(false));
-
-                if (emoteData == null)
-                    return;
-            }
-
-            JsonObject metadataJson = new JsonObject();
-            metadataJson.add("emote", emoteJson);
-            byte[] metadata = GSON.toJson(metadataJson).getBytes(StandardCharsets.UTF_8);
-
             try {
-                Files.write(emoteFile.toPath(), emoteData);
-                Files.write(metadataFile.toPath(), metadata);
+                Files.write(emoteFile.toPath(), dataMetadataPair.getKey());
+                Files.write(metadataFile.toPath(), dataMetadataPair.getValue());
             } catch (IOException e) {
                 LOGGER.fatal("Unable to write cache files.");
                 LOGGER.debug(e);
             }
-            LOGGER.info("Downloaded.");
+            LOGGER.debug("Downloaded emote.");
         } else {
-            LOGGER.info("Loaded from cache.");
+            LOGGER.debug("Loaded emote from cache.");
         }
 
         synchronized (LOCK) {
@@ -175,33 +116,5 @@ public class CachedEmoteProvider implements IEmoteProvider {
     @Override
     public boolean exists(long emoteID) {
         return available.contains(emoteID);
-    }
-
-    @Override
-    public boolean exists(String displayName) {
-        return displayNameToID.containsKey(displayName);
-    }
-
-    @Override
-    public long getEmoteID(String displayName) {
-        Long id = displayNameToID.get(displayName);
-        if (id == null)
-            throw new IllegalArgumentException("No such emote exists.");
-
-        return id;
-    }
-
-    @Override
-    public void requestFromServer(long emoteID) {
-        if (!requested.contains(emoteID)) {
-            requested.add(emoteID);
-
-            CraftCord.NETWORK.sendToServer(new RequestEmoteMessageC2S(emoteID));
-        }
-    }
-
-    @Override
-    public Set<String> usableEmotes() {
-        return Collections.unmodifiableSet(usable);
     }
 }
